Hompage: Simplify login/logout/switch accounts.
