+++ packages/user/Homepage/ui/src/hooks/use-connected-accounts.ts
@@ -2,11 +2,12 @@ import { useQuery } from "@tanstack/react-query";
 
 import { supervisor } from "@/supervisor";
 
+import QueryKey from "@/lib/queryKeys";
 import { zAccount } from "@/lib/zod/Account";
 
 export const useConnectedAccounts = () =>
     useQuery({
-        queryKey: ["connectedAccounts"],
+        queryKey: QueryKey.connectedAccounts(),
         initialData: [],
         queryFn: async () =>
             zAccount.array().parse(
