--- packages/user/Homepage/ui/src/hooks/use-select-account.ts
@@ -1,33 +1,0 @@
-import { queryClient } from "@/main";
-import { useMutation } from "@tanstack/react-query";
-
-import { supervisor } from "@/supervisor";
-
-import QueryKey from "@/lib/queryKeys";
-import { zAccount } from "@/lib/zod/Account";
-
-import { toast } from "@shared/shadcn/ui/sonner";
-
-import { setCurrentUser } from "./use-current-user";
-
-export const useSelectAccount = () =>
-    useMutation<void, Error, string>({
-        mutationKey: ["selectAccount"],
-        mutationFn: async (accountName: string) => {
-            void (await supervisor.functionCall({
-                method: "login",
-                params: [zAccount.parse(accountName)],
-                service: "accounts",
-                intf: "activeApp",
-            }));
-        },
-        onSuccess: (_, accountName) => {
-            setCurrentUser(accountName);
-            queryClient.refetchQueries({
-                queryKey: QueryKey.currentUser(),
-            });
-        },
-        onError: (error) => {
-            toast.error(error.message);
-        },
-    });
