+++ packages/user/Homepage/ui/src/layout.tsx
@@ -1,7 +1,10 @@
 import { Outlet, useNavigate } from "react-router-dom";
 
 import { AppSidebar } from "@/components/app-sidebar";
+import { AppSplashScreen } from "@/components/app-splash-screen";
+import { Loading } from "@/components/loading";
 
+import { useCurrentUser } from "@/hooks/use-current-user";
 import { useNavLocation } from "@/hooks/use-nav-location";
 
 import {
@@ -12,52 +15,12 @@ import {
     BreadcrumbPage,
     BreadcrumbSeparator,
 } from "@shared/shadcn/ui/breadcrumb";
-import {
-    Card,
-    CardDescription,
-    CardFooter,
-    CardHeader,
-    CardTitle,
-} from "@shared/shadcn/ui/card";
 import { Separator } from "@shared/shadcn/ui/separator";
 import {
     SidebarInset,
     SidebarProvider,
     SidebarTrigger,
 } from "@shared/shadcn/ui/sidebar";
-
-import { Loading } from "./components/loading";
-import { useCurrentUser } from "./hooks/use-current-user";
-import { LoginButton } from "./login-button";
-
-const SplashScreen = () => {
-    const { currentApp } = useNavLocation();
-
-    return (
-        <div className="mx-auto mt-4 w-[350px]">
-            {/* Main app info card */}
-            <Card className="rounded-b-none border-b-0 shadow-sm">
-                <CardHeader>
-                    <div className="mx-auto">{currentApp?.icon}</div>
-                    <CardTitle>{currentApp?.name}</CardTitle>
-                    <CardDescription>{currentApp?.description}</CardDescription>
-                </CardHeader>
-            </Card>
-
-            {/* Login prompt card */}
-            <Card className="bg-muted/50 rounded-t-none border-t-0">
-                <CardHeader className="pb-2 pt-4">
-                    <CardDescription className="text-center font-medium">
-                        {`Please log in to access ${currentApp?.name}`}
-                    </CardDescription>
-                </CardHeader>
-                <CardFooter className="flex justify-center pb-4">
-                    <LoginButton />
-                </CardFooter>
-            </Card>
-        </div>
-    );
-};
 
 export const Layout = () => {
     const navigate = useNavigate();
@@ -116,7 +79,7 @@ export const Layout = () => {
                 {isLoginRequired ? (
                     <div className="flex flex-1 items-center justify-center">
                         <div className="w-full max-w-screen-sm px-4">
-                            <SplashScreen />
+                            <AppSplashScreen />
                         </div>
                     </div>
                 ) : isPendingCurrentUser &&
