+++ packages/user/Homepage/ui/src/main.tsx
@@ -2,8 +2,7 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import ReactDOM from "react-dom/client";
 import { RouterProvider } from "react-router-dom";
 
-import { ThemeProvider } from "@/components/theme-provider";
-
+import { ThemeProvider } from "@shared/components/theme-provider";
 import { Toaster } from "@shared/shadcn/ui/sonner";
 
 import router from "./router";
